Create RiemannianGradientOptimizer
